-Add destroy()

+++ kublas.h
@@ -7,7 +7,7 @@ extern "C" {
 #endif
 	JNIEXPORT jlong JNICALL Java_kuda_Kublas_create(JNIEnv* env, jobject obj);
 
-	JNIEXPORT jlong JNICALL Java_kuda_Kublas_destroy(JNIEnv* env, jobject obj, jlong handle);
+	JNIEXPORT jint JNICALL Java_kuda_Kublas_destroy(JNIEnv* env, jobject obj, jlong handle);
 
 #ifdef __cplusplus
 }
